// src/app/shared/models/document-information/document-information.model.ts
export class DocumentInfoResponseModel {
  success: boolean;
  message: string;
  clientCode: string;
  clientType: string;
  domicileCondition: string;
  contractorStatus: string;
  isProspect: boolean;
  isState: boolean;
  promotions: boolean;
  personType: number;
  documentType: number;
  documentNumber: string;
  legalName: string;
  clienteAnulado: boolean;
  clienteEstado: boolean;
  names: string;
  apePat: string;
  apeMat: string;
  birthdate: string;
  returnBirthdate: boolean;
  image: string;
  nationality: number;
  department: number;
  province: number;
  district: number;
  email: string;
  phoneNumber: number;
  civilStatus: number;
  title: string;
  address: string;
  sex: number;
  idSeguridad: string;
  id: string;

  constructor(payload?: any) {
    this.clienteAnulado = payload?.ClienteAnulado || null;
    this.clienteEstado = payload?.ClienteEstado || null;

    const locationCodeExcludes = [3, 33];

    this.success = payload?.Success || false;
    this.message = payload?.Message || null;

    this.clientCode = payload?.CodigoCliente || null;
    this.clientType = payload?.TipoCliente || null;
    this.domicileCondition = (
      (payload?.CondicionDomicilio || '') as string
    ).toUpperCase();
    this.contractorStatus = (
      (payload?.EstadoContratante || '') as string
    ).toUpperCase();
    this.isProspect = payload?.ClienteProspecto || false;
    this.isState = payload?.ClienteEstado || false;
    this.promotions = payload?.Promociones || false;
    this.personType = +payload?.IdTipoPersona || null;
    this.documentType = +payload?.IdTipoDocumento || null;
    this.documentNumber = payload?.NumeroDocumento || null;
    this.legalName =
      +payload?.IdTipoDocumento == 1 ? payload?.RazonSocial || null : null;
    this.names = payload?.Nombre || null;
    this.apePat = payload?.ApellidoPaterno || null;
    this.apeMat = payload?.ApellidoMaterno || null;

    this.image = payload?.Imagen || null;
    // tslint:disable-next-line:max-line-length
    this.birthdate = payload?.FechaNacimiento || null;
    this.returnBirthdate = !!payload?.FechaNacimiento;

    this.nationality =
      +payload?.IdTipoDocumento == 2 ? 1 : +payload?.Nacionalidad || null;

    this.department = +payload?.Departamento || null;
    this.province = +payload?.Provincia || null;
    this.district = +payload?.Distrito || null;

    const locations = [this.department, this.province, this.district];
    if (locations.every((x) => x == this.department)) {
      this.department = null;
      this.province = null;
      this.district = null;
    }

    this.address = payload?.Direccion || null;

    this.email = payload?.Email || null;
    this.phoneNumber = +payload?.Telefono || null;

    this.civilStatus = +payload?.EstadoCivil || null;
    this.title = payload?.Titulo || null;
    this.sex = +payload?.Sexo || null;
    this.idSeguridad = payload?.IdSeguridad || null;
    this.id = payload?.Id || null;
  }
}